IBM/requirements.txt
asgiref==3.3.4
boto3==1.17.101
botocore==1.20.101
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
colorama==0.4.4
Django==3.2.4
django-cors-headers==3.7.0
django-rest-framework==0.1.0
djangorestframework==3.12.4
filelock==3.0.12
huggingface-hub==0.0.12
idna==2.10
jmespath==0.10.0
joblib==1.0.1
numpy==1.21.0
packaging==20.9
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
regex==2021.4.4
requests==2.25.1
s3transfer==0.4.2
sacremoses==0.0.45
semantic-version==2.8.5
sentencepiece==0.1.96
setuptools-rust==0.12.1
six==1.16.0
sqlparse==0.4.1
tokenizers==0.5.2
toml==0.10.2
torch==1.9.0
tqdm==4.61.1
transformers==2.8.0
typing-extensions==3.10.0.0
urllib3==1.26.6
